requirements.txt
streamlit
google-generativeai
pandas
plotly
python-dotenv
fpdf2
